python/requirements.txt
opencv-python
pytesseract
Pillow
torch
torchvision
Flask
Werkzeug
python-dotenv
pandas
